gke_upgrade_tool/requirements.txt
Requests==2.31.0
ruamel.yaml==0.18.5